feat: Migrate from GDrive to AWS feat: Migrate to New Kotlin Multiplatform AGP feat: Migrate to Kotlin 2.3.0 bug fixes remove baseline profile module cleanups

+++ feature/map/src/commonMain/kotlin/com/rafael/inclusimap/feature/map/map/presentation/components/FindNorthWidget.kt
@@ -50,7 +50,7 @@ fun BoxScope.FindNorthWidget(
                 MaterialTheme.colorScheme.error
             },
             modifier = Modifier
-                .rotate(degrees = cameraPosition.bearing.toFloat()),
+                .rotate(degrees = cameraPosition.bearing),
         )
     }
 }
